detection-service/requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
ultralytics>=8.0.0  # For YOLO models
supervision>=0.3.0  # For object tracking 